Move some blocker ui elements
