+++ app/src/main/java/com/undistract/ui/blocking/BlockedAppActivity.kt
@@ -1,10 +1,9 @@
-package com.undistract
+package com.undistract.ui.blocking
 
 import android.os.Bundle
 import android.os.Handler
 import android.os.Looper
 import androidx.appcompat.app.AppCompatActivity
-//import android.content.Intent
 
 class BlockedAppActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -16,15 +15,6 @@ class BlockedAppActivity : AppCompatActivity() {
         // Close this activity after a delay
         Handler(Looper.getMainLooper()).postDelayed({
             // Create an intent to launch MainActivity
-//            val intent = Intent(this, MainActivity::class.java).apply {
-//                // Clear the back stack so user can't return to blocked app
-//                flags = Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_CLEAR_TASK
-//
-//                // Optional: Pass information about the blocked app
-//                putExtra("came_from_blocked_app", true)
-//                putExtra("blocked_app_name", appName)
-//            }
-//            startActivity(intent)
             finish()
         }, 500)
     }
